+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [Alpha 0.2.4] 2025-02-02
+
+### Added
+- Column "month" to requestquery.
+
+### Changed
+- dayoffinfo now auto detect the year of the ticket when a new request is made.
+
+---
 ## [Alpha 0.2.3] 2025-01-31
 
 ### Changed